+++ android/src/main/kotlin/com/dns_technologies/mlkit_scanner/CameraViewScannerCamera.kt
@@ -70,6 +70,7 @@ class CameraViewScannerCamera(private val lifecycleOwner: LifecycleOwner,
     }
 
     override fun attachAnalyser(analyzer: CameraImageAnalyzer) {
+        clearAnalyzer()
         this.analyzer = analyzer
         cameraView.addFrameProcessor(this::analyzeFrame)
     }