+++ android/src/main/kotlin/com/dns_technologies/mlkit_scanner/MlkitScannerPlugin.kt
@@ -26,7 +26,6 @@ import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
-import io.flutter.plugin.common.PluginRegistry.Registrar
 import kotlin.Exception
 
 class PermissionsConstants {
@@ -80,7 +79,7 @@ class MlkitScannerPlugin: FlutterPlugin, MethodCallHandler, ActivityAware, Lifec
       PluginConstants.setScanDelayMethod -> invokeSetScanDelay(call, result)
       PluginConstants.updateConstraintsMethod -> result.success(true) // на Android нет необходимости обрабатывать
       PluginConstants.setZoomMethod -> invokeSetZoom(call, result)
-      PluginConstants.addCropOverlayMethod -> invokeAddCropOverlay(call, result)
+      PluginConstants.setCropAreaMethod -> invokeSetCropArea(call, result)
       else -> result.notImplemented()
     }
   }
@@ -161,17 +160,21 @@ class MlkitScannerPlugin: FlutterPlugin, MethodCallHandler, ActivityAware, Lifec
 
   private fun invokeStartScan(call: MethodCall, result: Result) {
     val options = AnalyzeOptions.fromMap(call.arguments as Map<String, Any?>)
-    analyzer?.resumeScan(options.periodMs)
-    if(checkCameraActiveStatus(result,
-        "You need to invoke \'initCameraPreview\' method before start scan")) {
-      if (analyzer?.isDisposed == true || analyzer?.type != options.recognizeType) {
-        analyzer = AnalyzerCreator.create(options.recognizeType)
-        analyzer?.init(options.periodMs, this::onScan, cameraImagePreparer::prepare)
-      }
-      camera?.attachAnalyser(analyzer!!)
-      scannerOverlay?.isActive = true
-      result.success(true)
+    if (cameraLifecycle == null) {
+      return result.error(
+        PluginError.CameraIsNotInitialized.errorCode,
+        "You need to invoke \'initCameraPreview\' method before start scan",
+        null
+      )
     }
+    analyzer?.resumeScan(options.periodMs)
+    if (analyzer?.isDisposed == true || analyzer?.type != options.recognizeType) {
+      analyzer = AnalyzerCreator.create(options.recognizeType)
+      analyzer?.init(options.periodMs, this::onScan, cameraImagePreparer::prepare)
+    }
+    camera?.attachAnalyser(analyzer!!)
+    scannerOverlay?.isActive = true
+    result.success(true)
   }
 
   private fun invokeCancelScan(result: Result) {
@@ -204,6 +207,12 @@ class MlkitScannerPlugin: FlutterPlugin, MethodCallHandler, ActivityAware, Lifec
     cameraLifecycle = CameraLifecycle()
     createScannerCamera()
     cameraLifecycle!!.resume()
+    cameraView.addOnLayoutChangeListener { _, _, _, _, _, _, _, _, _ ->
+      val cropRect = scannerOverlay?.cropRect
+      if (cropRect != null) {
+        updateCropOptions(cropRect)
+      }
+    }
   }
 
   private fun createScannerCamera() {
@@ -249,11 +258,15 @@ class MlkitScannerPlugin: FlutterPlugin, MethodCallHandler, ActivityAware, Lifec
     }
   }
 
-  private fun invokeAddCropOverlay(call: MethodCall, result: Result) {
+  private fun invokeSetCropArea(call: MethodCall, result: Result) {
     val rect = RecognizeVisorCropRect.fromMap(call.arguments as Map<String, Any?>)
     updateCropOptions(rect)
-    scannerOverlay = ScannerOverlay(rect, cameraView.context)
-    cameraView.addOverlay(scannerOverlay!!)
+    if (scannerOverlay != null) {
+      scannerOverlay?.cropRect = rect
+    } else {
+      scannerOverlay = ScannerOverlay(rect, cameraView.context)
+      cameraView.addOverlay(scannerOverlay!!)
+    }
     result.success(true)
   }