+++ lib/widgets/camera_preview.dart
@@ -1,7 +1,9 @@
 import 'dart:async';
 import 'dart:io';
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:mlkit_scanner/platform/ml_kit_channel.dart';
 
@@ -13,15 +15,16 @@ typedef CameraInitilizeError = void Function(PlatformException);
 /// Widget automatically will dispose camera when called [dispose] in state.
 class CameraPreview extends StatefulWidget {
   /// Callback when device camera initialize.
-  final VoidCallback? onCameraInitialized;
+  final VoidCallback onCameraInitialized;
 
   /// Callback if camera cannot be initialized.
   final CameraInitilizeError? onCameraInitializeError;
 
-  CameraPreview({
-    this.onCameraInitialized,
+  const CameraPreview({
+    Key? key,
+    required this.onCameraInitialized,
     this.onCameraInitializeError,
-  });
+  }) : super(key: key);
 
   @override
   _CameraPreviewState createState() => _CameraPreviewState();
@@ -45,20 +48,38 @@ class _CameraPreviewState extends State<CameraPreview> {
           return UiKitView(
             viewType: 'mlkit/camera_preview',
             onPlatformViewCreated: _onViewCreated,
-            creationParamsCodec: StandardMessageCodec(),
+            creationParamsCodec: const StandardMessageCodec(),
             creationParams: {
               'width': constraints.maxWidth,
               'height': constraints.maxHeight,
             },
           );
         }
-        return AndroidView(
+        return PlatformViewLink(
           viewType: 'mlkit/camera_preview',
-          onPlatformViewCreated: _onViewCreated,
-          creationParamsCodec: StandardMessageCodec(),
-          creationParams: {
-            'width': constraints.maxWidth,
-            'height': constraints.maxHeight,
+          surfaceFactory: (context, controller) {
+            return AndroidViewSurface(
+              controller: controller as AndroidViewController,
+              gestureRecognizers: const {},
+              hitTestBehavior: PlatformViewHitTestBehavior.opaque,
+            );
+          },
+          onCreatePlatformView: (params) {
+            return PlatformViewsService.initSurfaceAndroidView(
+              id: params.id,
+              viewType: 'mlkit/camera_preview',
+              layoutDirection: TextDirection.ltr,
+              creationParams: {
+                'width': constraints.maxWidth,
+                'height': constraints.maxHeight,
+              },
+              creationParamsCodec: const StandardMessageCodec(),
+            )
+              ..addOnPlatformViewCreatedListener((id) {
+                params.onPlatformViewCreated(id);
+                _onViewCreated(id);
+              })
+              ..create();
           },
         );
       },
@@ -74,7 +95,7 @@ class _CameraPreviewState extends State<CameraPreview> {
   Future<void> _onViewCreated(int id) async {
     try {
       await _channel.initCameraPreview();
-      widget.onCameraInitialized?.call();
+      widget.onCameraInitialized();
     } on PlatformException catch (e) {
       widget.onCameraInitializeError?.call(e);
     }