+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.2.0
+
+* Migrate from AndroidView to PlatformViewLink
+
+New Features
+
+* resize CameraPreview widget at runtime
+* change crop area at runtime
+
 ## 0.1.0
 
 * Initial Release